Finished FBOs
I had to change the target in the tsconfig from es5 to es6 in order to
use a certain function and it gave me errors which I fixed in the file
lib.es6.d.  If it gives you errors about duplicates just replace the
file in rectrix-app/node_modules/typescript/lib with the version I moved
into /rectrix-app and it should fix it.